refactor: move generate_default_file_path to utils.py for better maintainability
